requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
cachetools==7.0.5
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
cryptography==46.0.6
cymem==2.0.13
dataclasses-json==0.6.7
ddgs==9.12.0
distro==1.9.0
duckduckgo_search==8.1.1
emoji==2.15.0
en_core_web_sm==3.8.0
faiss-cpu==1.13.2
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.69.0
greenlet==3.3.2
groq==0.37.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
installer==0.7.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
llvmlite==0.46.0
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.18.1
numba==0.64.0
numpy==2.4.4
olefile==0.47
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
preshed==3.0.13
primp==1.2.1
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pypdf==6.9.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-magic-bin==0.4.14
python-oxmsg==0.0.2
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.3
streamlit==1.55.0
tenacity==9.1.4
thinc==8.3.13
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.22.10
unstructured-client==0.42.12
urllib3==2.6.3
uuid_utils==0.14.1
wasabi==1.1.3
watchdog==6.0.0
weasel==1.0.0
webencodings==0.5.1
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
